+++ CHANGELOG.md
@@ -2,6 +2,30 @@
 
 ## Next Release
 
+## 0.0.56
+
+- [subgraph] Use bware subgraph endpoints
+
+## 0.0.55
+
+- [sdk] Fix lower bound
+
+## 0.0.54
+
+- [contracts] Update default subgraphs
+
+## 0.0.53
+
+- [contracts] Redeploy to mainnets
+- [utils] Update default subgraph URLs
+
+## 0.0.52
+
+- [sdk] Add sanity check for callTo parameter
+- [sdk] update sdk constructor params structure again
+- [router] perf subgraph buffer
+- [contracts] Redeploy contracts to testnets
+
 ## 0.0.51
 
 - [contracts] Final updates from auditor